% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_planaflor_data_dir.R
\name{get_planaflor_data_dir}
\alias{get_planaflor_data_dir}
\title{Check/get the data folder for the PlanaFlor project}
\usage{
get_planaflor_data_dir()
}
\value{
a path to the data folder of the PlanaFlor project
}
\description{
Check/get the data folder for the PlanaFlor project
}
\details{
The \code{get_planaflor_data_dir()} function check the
existence of the enviromental variable PLANAFLOR on the global
\code{.Renviron} file, and offer instructions on how to set that
variable. The main objective of this function is to offer a
concise way to get to planaflor data irrespective the computer
which the data will be accessed.
}
\examples{
\dontrun{
if (interactive()) {
  get_planaflor_data_dir()
  # If you get an error, don't worry.
  # Follow the instructions provided by the error message
  # and rerun get_planaflor_data_dir.
}
}

}
\seealso{
\code{\link[glue]{glue}}
\code{\link[usethis]{ui}}
\code{\link[here]{here}}
}
